Première version non testée

+++ 7emeArcheLibrairie/src/main/java/com/online/eLibrairie/repositories/AdresseRepository.java
@@ -1,10 +1,15 @@
 package com.online.eLibrairie.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.online.eLibrairie.models.Adresse;
-
-public interface AdresseRepository extends JpaRepository<Adresse, Integer>{
+/**
+ * Interface for generic CRUD operations on a repository for user's adresse
+ * we can implement certain methods for managing adresse of the user
+ * like findAll(),findById(),save()
+ * @author stella
+ *
+ */
+public interface AdresseRepository extends JpaRepository<Adresse, Long>{
 	
 
 }
